requirements.txt
PyYAML==6.0.1
types-PyYAML==6.0.12.20241230